refactor: change api folder directories
