+++ renderer/src/apis/repo/apis.ts
@@ -1,0 +1,7 @@
+import { API } from '..';
+
+const getRepoList = async (userId: string, sortBy: string) => {
+  return API.get(`/users/${userId}/repos?sort=${sortBy}`);
+};
+
+export { getRepoList };
