+++ renderer/src/apis/repo/index.ts
@@ -1,7 +1,2 @@
-import { API } from '..';
-
-const getRepoList = async (userId: string, sortBy: string) => {
-  return API.get(`/users/${userId}/repos?sort=${sortBy}`);
-};
-
-export { getRepoList };
+export * from './apis';
+export * from './types';
