+++ renderer/src/apis/repo/types.ts
@@ -1,0 +1,6 @@
+export const RepoSortBy = {
+  CREATED: 'created',
+  UPDATED: 'updated',
+  PUSHED: 'pushed',
+  FULL_NAME: 'full_name',
+} as const;
